requirements.txt
discord.py[voice]
asyncpg
wavelink
humanize
wikia
git+https://github.com/Rapptz/discord-ext-menus
jishaku
parsedatetime